==> notebooks/02_data_cleaning.ipynb <==
from pprint import pprint
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings("ignore")

# loading the scraped dataset
df = pd.read_csv("../data/reddit_tech_raw_data.csv")
df.head()

df.shape

# creating a copy of dataset
df_copy = df.copy()

df_copy.shape

df_copy.info()

# fixing the timestamp format
df_copy['timestamp'] = pd.to_datetime(df_copy['timestamp'], unit = 's')
df_copy.head()

df_copy.info()

# checking for null values
missing_count = df_copy.isnull().sum().sort_values(ascending = False)
missing_count

#### About Null Values
- `title`and `post_url`columns are only relevant to `posts` since the majority of rows are comments(over 13000) these columns are empty
- `text` column is mainly for the `comments` like body of the comments. Posts may have text if they are self-posts but most of the posts contain a title and a link so its showing 100 null values

**Action:**
- Filling `NaN` values in `text` and `title` columns with empty strings

# filling NaN values with the empty strings
df_copy['title'] = df_copy['title'].fillna(" ")
df_copy['text'] = df_copy['text'].fillna(" ")
df_copy['post_url'] = df_copy['post_url'].fillna(" ")

# for sentiment analysis creating a new column
df_copy['text_to_analyze'] = np.where(df_copy['type'] == 'post',
                                      df_copy['title'] + " " + df_copy['text'],
                                      df_copy['text'])

df_copy[['type', 'title', 'text', 'text_to_analyze']].head()

# looking for duplicates
print(f"Number of Duplicate rows: {df_copy.duplicated().sum()}")

# checking index values are unique or not
df_copy.index.is_unique

# checking for unique values in categorical columns
print(f"Unique Values in 'type' column: {df_copy['type'].unique()}") # expected 2 unique values 'post' and 'column'
print(f"Unique Values in 'post_id' column: {df_copy['post_id'].unique()}") # scraped 100 posts and output displayed 100 unique values

### Text Preprocessing Step
- converting all the text to lowercase on the `text_to_analyze` column
- removing any urls, emojis, punctuation and special characters
- removing any stop words that appear repeatedly but doesn't add any value to the sentence
- splitting the body of words to tokens and grouping the similar words and reduce the size of the vocabulary by stemming process

import re
import nltk
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer

nltk.download('stopwords')
nltk.download('punkt')

stemmer = PorterStemmer()
stop_words = set(stopwords.words('english'))

# creating a function to clean the text_to_analyze column
def clean_text(text):
  text = text.lower()   # converting all the text to lower case
  text = re.sub(r'http\S+|www\S+|@\w+|#\w+', '', text)  # removes url, mentions and hashtags
  text = re.sub(r'[^a-z\s]', '', text)  # removes all the punctuation and number only keeps the letters and spaces
  text = re.sub(r'\s+', ' ', text).strip()   # removes the extra whitespaces

  tokens = text.split()   # this breaks text to tokens
  filtered_stemmed_tokens = [
    stemmer.stem(word) for word in tokens if word not in stop_words
  ]
  cleaned_text = ' '.join(filtered_stemmed_tokens)    # this will join the filtered tokens back to single string

  return cleaned_text

df_copy['cleaned_text'] = df_copy['text_to_analyze'].apply(clean_text)

print(df_copy[['text_to_analyze', 'cleaned_text']].head(10))

df_copy.to_csv('../data/reddit_tech_cleaned_data.csv', index = False)

#### moving to next step i.e., EDA on SQL(MySQL)
