==> notebooks/03_02_eda.ipynb <==
import pandas as pd
import mysql.connector
from dotenv import load_dotenv
import os
from urllib.parse import quote_plus
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from collections import Counter

load_dotenv()

# loading the cleaned data
df = pd.read_csv('../data/cleaned_data.csv')

df_copy = df.copy()

# Filling NaN values with None so SQL can read
df_copy = df_copy.replace({pd.NA: None, pd.NaT: None, float('nan'): None})

# database connection
db_user = os.getenv('DB_USER')
db_password = os.getenv('DB_PASSWORD')
db_host = os.getenv('DB_HOST')
db_name = os.getenv('DB_NAME')

password_encoded = quote_plus(db_password)

# connecting to MySQL database
conn = mysql.connector.connect(
  host=db_host,
  user=db_user,
  password=db_password,
  database=db_name
)

# converting dataframe to list of tuples to insert into SQL
data_tuples = [tuple(x) for x in df_copy.to_numpy()]

cursor = conn.cursor()
sql = "INSERT INTO post_and_comment (type, post_id, title, timestamp, text, score, total_comments, post_url, text_to_analyze, cleaned_text) VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)"

cursor.executemany(sql, data_tuples)
conn.commit()
cursor.close()
conn.close()

print("Data Successfully loaded into MySQL table.")
