==> notebooks/04_sentiment_analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import nltk
from nltk.sentiment import SentimentIntensityAnalyzer
from nrclex import NRCLex

# loading the dataset from the previous step
df = pd.read_csv("../data/processed_reddit_data.csv")

df_copy = df.copy()
df_copy.head()

#### starting with the 'Sentiment Analysis' using VADER which is a part of `nltk` library

nltk.download('vader_lexicon')

# initializing the VADER sentiment analyzer
analyzer = SentimentIntensityAnalyzer()

# function to get the sentiment score
def get_sentiment_score(text):
  if not isinstance(text, str):
    return None
  return analyzer.polarity_scores(text)['compound']

df_copy['sentiment_score'] = df_copy['cleaned_text'].apply(get_sentiment_score)

print(df_copy.head(10))

**Analysis:**
- row number 0, 1, 4, 7, 8 have positive score which indicates the positive tone
- row number 2, 6, 9 have negative score indicates negative tone
- rows with a score 0.0 indicates the netural tone

# grouping by 'topic' and calculating the average sentiment score
topic_sentiment = df_copy.groupby('topic')['sentiment_score'].mean().reset_index()
print(topic_sentiment)

**Analysis:**
- From the sentiment score, the `AI Related` and `AI and Privacy Related` topics have positive average sentiment scores `0.04`. This suggests community towards AI is slightly positive
- For `Privacy and Political Related` topics have the most negative scores `-0.12`. This shows the discussions about politics and privacy related in the `r/technology` subreddit are often critical or contentious
- But when its only about `Privacy Related` the tone is slightly negative to neutral.

##### **Improvising the project by adding some more emotions. I have already used the VADER so adding `NRC Emotion Lexicon` for this emotion classification**

nltk.download('punkt')

def get_emotions(text):
  if not isinstance(text, str):
    return
  emotion = NRCLex(text)
  return emotion.raw_emotion_scores

df_copy['emotions'] = df_copy['cleaned_text'].apply(get_emotions)

emotion_df = df_copy['emotions'].apply(pd.Series).fillna(0)

df_copy = pd.concat([df_copy, emotion_df], axis = 1)
print(df_copy.head())
